modify and refactor filter
