+++ services/unit.js
@@ -4,29 +4,15 @@ require('../models/Unit');
 const Unit = mongoose.model('Unit');
 const paginate = require('../helpers/paginate');
 const custom = require('../helpers/custom');
+const filters = require('../helpers/filter/unitfilter');
 
 const listUnit = async (query, callback) => {
     try {
         const sorts = (query.sort == "desc" ? { createdAt: "desc" } : JSON.parse(query.sort));
         const populate = (['createdBy']);
         const options = paginate.optionsLabel(query, sorts, populate);
-        let params = {};
-        if(query.unit_type){
-            params.unit_type = query.unit_type;
-        }
-        if(query.code_district_code){
-            params.code_district_code = query.code_district_code;
-        }
-        let search_params;
-        if(query.search){ 
-            search_params = [
-                { unit_code: new RegExp(query.search, "i") },
-                { unit_type: new RegExp(query.search, "i") },
-                { name: new RegExp(query.search, "i") }
-            ];
-        }else{
-            search_params = {};
-        }
+        const params = filters.filterUnit(query);
+        const search_params = filters.filterSearch(query);
         const result = Unit.find(params).or(search_params).where('delete_status').ne('deleted');
         const paginateResult = await Unit.paginate(result, options);
         callback(null, paginateResult);
@@ -64,7 +50,7 @@ const updateUnit = async (pay, id, category, author, callback) => {
         const payloads = {};
         const payload = (pay == null ? {} : pay);
         if (category == "delete") {
-            custom.deletedSave(payloads);
+            custom.deletedSave(payloads, author);
         }
         const params = Object.assign(payload, payloads);
         const result = await Unit.findByIdAndUpdate(id,
